Remove not-working Kiln test (#34)
At least 2 problems in the feature file (unmodified
`your_interview.yml` as the test yaml, and an older
`acknowledged_information_use['I agree to the terms']`
variable that doesn't exist anymore), so IMO it's more
worth it to remove and rebuild trust in the tests rather
than doing a deeper dive (when I'm still running into issues
running Kiln locally).

It'll still be in the history (and I strongly suspect that
the test came from https://plocket.github.io/alkiln_story/
without any testing or modification, so it's probably a better
idea to redo that process).